cloud-run/demo4/requirements.txt
Flask==2.3.2
google-auth==2.14.1
google-auth-oauthlib==0.4.6
PyJWT